backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart
pydantic
python-dotenv
itsdangerous
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-generativeai
google-genai
reportlab
weasyprint
Jinja2